Add Link To Image In Docker Hub
The point of this project is to generate a usable Docker Image from
which people can spawn Docker Containers to use when building GCC based
code projects that need to include the cpputest library.  Put a link to
the Docker Image in Docker Hub and give the docker pull command.